Latest Commit: Moved the 'Sphere' class implementation from 'Scene.h' to it's own 'Sphere.h' and 'Sphere.cpp' file
